Refactored the undo-action to be inside the command
This way the logic on how to undo an action is in the command itself, as
opposed to spread around the editor, and is set when the command is
created, at the point where the information is freshest and most
relevant (i.e. we don't have to do any magic on the arguments to get the
results that we wanted).

This way too, one just has to call the function instead of having to
remember that additional logic needs to be put in other areas of the
code.

This also makes it trivial to add another command type